backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic
pydantic-settings
faiss-cpu 
sentence-transformers
google-generativeai
python-jose 
bcrypt
redis 
rq
python-multipart
pgvector